Bastantes cosillas

+++ client/src/components/Profile/Profile.tsx
@@ -14,6 +14,7 @@ import Button from '@mui/material/Button';
 import PostAPet from "../PostAPet"
 import { useSelector, useDispatch } from 'react-redux';
 import { typeState } from '../../redux/reducers';
+import Swal from 'sweetalert2';
 
 
 
@@ -46,8 +47,42 @@ export default function Profile() {
   const id = localStorage.getItem('userId');
 
   const handleClick = (id: string | undefined) => {
-    deletePostService(id);
-    setPosts(posts.filter(post => post._id !== id));
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: '¿Seguro que quieres eliminar la publicacion?',
+      text: "No podras revertir esta accion",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Si, quiero eliminarla!',
+      cancelButtonText: 'No, cancelar!',
+      reverseButtons: false
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deletePostService(id);
+        setPosts(posts.filter(post => post._id !== id));
+        swalWithBootstrapButtons.fire(
+          'Eliminada!',
+          'Tu publicacion ha sido eliminada',
+          'success'
+        )
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelado',
+          'Tu publicacion NO se ha eliminado',
+          'error'
+        )
+      }
+    })
   };
 
   // let refresh = useSelector((state: typeState) => state.editPost);
